:recycle: Separate App.tsx
